Move Stripe preload to OnboardingStepLayout and extract shared helpers
Replace the two per-page preload calls with a single call in the shared
onboarding step layout, extract the plan-required predicate into
getIsPlanRequired for reuse by useSetNextOnboardingStatus, move the
promise cache from the useStripePromise hook file to a billing util,
and drop the useEffect in favor of the render-time call idiom the
sibling hook already uses.
