+++ packages/twenty-front/src/modules/onboarding/utils/getIsPlanRequired.ts
@@ -1,0 +1,11 @@
+import { type CurrentWorkspace } from '@/auth/states/currentWorkspaceState';
+
+export const getIsPlanRequired = ({
+  isBillingEnabled,
+  currentWorkspace,
+}: {
+  isBillingEnabled: boolean;
+  currentWorkspace: Pick<CurrentWorkspace, 'billingSubscriptions'> | null;
+}) =>
+  isBillingEnabled &&
+  (currentWorkspace?.billingSubscriptions?.length ?? 0) === 0;
