+++ packages/twenty-front/src/modules/settings/billing/utils/__tests__/getStripePromise.test.ts
@@ -1,6 +1,6 @@
 import { loadStripe } from '@stripe/stripe-js/pure';
 
-import { getStripePromise } from '@/settings/billing/hooks/useStripePromise';
+import { getStripePromise } from '@/settings/billing/utils/getStripePromise';
 
 jest.mock('@stripe/stripe-js/pure', () => ({
   loadStripe: jest.fn(),
